Logs/metagpt/part3-2/instruction31/azure_chatbot_backend/backend/requirements.txt
fastapi==0.110.2
httpx==0.27.0
python-dotenv==1.0.1
uvicorn==0.29.0
gunicorn==21.2.0